requirements.txt
numpy
jax
jaxlib
pyyaml
